move filter_mp functions out
